Rename utils under __tests__ to test-utils
