+++ src/lazy/__tests__/lazy-server.unit.test.tsx
@@ -12,7 +12,7 @@ import {
   testFallbackRender,
   testRender,
   TestRenderOptions,
-} from './utils';
+} from './test-utils';
 
 jest.mock('../../utils', () => ({
   ...jest.requireActual<any>('../../utils'),
